pyproject.toml
[project]
name = "mcp-fmi"
dynamic = ["version"]
description = "Unofficial integration of fmpy with mcp."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dash>=3.0.4",
    "mcp[cli]>=1.9.0",
    "python-dotenv>=1.1.0",
    "fmpy",
    "numpy",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "build",
    "twine",
    "hatch",
]

[project.scripts]
mcp-fmi = "mcp_fmi.server:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_fmi"]

[tool.hatch.version]
source = "vcs"
